backend/back_end/requirements.txt
anyio==4.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.35.20
botocore==1.35.20
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels-redis==4.2.1
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
defusedxml==0.8.0rc2
Django==4.2.16
django-allauth==65.0.2
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==4.4.0
django-rest-auth==0.9.5
django-ses==4.1.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
exceptiongroup==1.2.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rpds-py==0.21.0
s3transfer==0.10.2
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.1.0
twisted==24.10.0
txaio==23.1.1
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
zope.interface==7.1.1
